Dl/requirements.txt
fastapi
uvicorn==0.27.1
python-multipart
numpy
torch
torchvision
Pillow
python-consul
google-generativeai
python-dotenv
